Rename get-size to getSize.
