+++ lib/Model.js
@@ -17,7 +17,6 @@ var arraySlice = require("./support/array-slice");
 var collectLru = require("./lru/collect");
 var pathSyntax = require("falcor-path-syntax");
 
-var getSize = require("./support/get-size");
 var isObject = require("./support/is-object");
 var isFunction = require("./support/is-function");
 var isPrimitive = require("./support/is-primitive");
@@ -26,6 +25,7 @@ var isJsonGraphEnvelope = require("./support/is-json-graph-envelope");
 
 var setCache = require("./set/set-cache");
 var setJsonGraphAsJsonDense = require("./set/set-json-graph-as-json-dense");
+var getSize = require("./support/getSize");
 var jsong = require("falcor-json-graph");
 var ID = 0;
 var validateInput = require("./support/validate-input");
