+++ lib/response/IdempotentResponse.js
@@ -5,7 +5,7 @@ var ModelResponse = require("./../response/ModelResponse");
 
 var pathSyntax = require("falcor-path-syntax");
 
-var getSize = require("./../support/get-size");
+var getSize = require("./../support/getSize");
 var collectLru = require("./../lru/collect");
 
 var arrayClone = require("./../support/array-clone");
