+++ lib/support/getSize.js
@@ -1,4 +1,4 @@
-var isObject = require("./../support/is-object");
+var isObject = require("./../support/isObject");
 module.exports = function getSize(node) {
     return isObject(node) && node.$size || 0;
 };
